Add more notes on interpreter behavior and specs
A file can be treated as a module or invoked as a script from `python'
based on the use of `-m' flag _without_ .py extension, or called as the
path to the file, respectively.

Updated the spec with a couple more section test cases. Deferring to
external tools for publish and search solves two issues:

  * We don't need subsections, just a long run of text with occasional
    breaks.
  * A search tool could provide in-context results when looking for text
    Like `apropos' or `man -k' but paragraph-aware.

It's still flexible if folks want to have a book be

  One long file with sections as chapters, index, and appendices.

  Or each file is a chapter with sections as subsections.

  Or each folder is a chapter and supporting files (graphics, LaTeX, and
  others) are woven together (by the separate _publishing_ tool).