Implement a line-counting parser w/ sanity handler
So, two things:

  New parser! It's "smarter" because it stores state and passes it on to
  the handler.

  Backward-incompatible sanity handler :( Because it takes in new params
  line count and processed line count.

This exactly reflects M.A. Jackson's admonishment not to use plain for
loops, an indicator of a poor understanding of the structure of the data
Fatal to a parser project -_-

Given the shortcomings, this is a demonstration of a parser that can
pass things to a handler.

A generic thing would be nice. Interface? Structs?