Add parser that handles state via hash object
We can pass state with a `{}' object instead of a long line of params
for the handler, genericizing the next line of handlers.

The new handler (will or is supposed to) stores the last line, so that
we can check whether to start or end lists, and maybe multiple sentences
in a paragraph versus one sentence as a paragraph.